review: HF-style model fixtures + test cleanups
Addresses PR review on the model fixtures and tests:
- models organized like HF modeling files: one component per file
  (attention/gdn/moe each its own file), a few @funcs compose the
  component Module, the top model Module includes the components;
  Qwen3.5 split out of the single qwen35_module.py into per-component
  files (shared dims in config.py)
- drop the retired v1/v2 attention placeholders and version suffixes
  (mla_kv_update_v2 -> mla_kv_update, mla_attend_v2 -> mla_attend)
- comments in the model files are English-only, structure-only
- test_nested_module uses ir.types.utils.make_tensor_type instead of a
  hand-rolled TensorType helper
- fold test_topk_dynamic_k.py into test_topk.py
