+++ tests/models/deepseek_v4_flash/attention.py
@@ -1,155 +1,30 @@
-"""DeepSeek V4 flash decode-step attention component.
+"""DeepSeek-V4-Flash decode-step attention component: real transformer layer 0
+(config.json ``compress_ratios[0] == 0``) -- pure sliding-window MLA.
 
-Head configuration is **GQA** (32 query heads / 4 KV heads / 128 head dim) as
-a placeholder for the real MLA attention — MLA's own head/latent dims land
-once the architecture parameters are confirmed; the leaf boundary this module
-exposes (``attn_kv_update`` / ``attn_scores``) does not change either way.
+Two chained ``@func``s (the caller makes two separate ``evaluate()`` calls,
+not one composed ``@func``): ``mla_kv_update`` projects and RMSNorms the
+shared KV latent, applies partial RoPE to its last ``REAL_ROPE_DIM`` dims,
+fp8 fake-quantizes the non-rope portion, and writes the result into a fixed
+``REAL_WINDOW``-token cache slot. ``mla_attend`` computes the low-rank Q
+projection (``wq_a`` -> RMS rescale -> ``wq_b``), applies the same partial
+RoPE, attends the single new-token query over the cached KV latent (MQA:
+``n_kv_heads == 1``, the cache serves as both K and V -- MLA-absorbed, no
+separate V projection) with an ``attn_sink`` softmax column, inverse-RoPEs
+the context, and applies the grouped low-rank O projection (``REAL_O_GROUPS``
+groups, ``wo_a`` per group -> ``wo_b``).
 
-Mirrors ``tests/models/qwen3_5_30b_a3b/common.py``'s ``build_kv_update`` /
-``build_scores`` split: a concat-derived compound ``DimVar`` axis (prior +
-new context length) cannot feed ``matmul`` / ``transpose`` in the same
-Function (the relation engine cannot recover it), so the KV-cache append and
-the score computation are two ``@func``s, chained in Python — see
-``tests/models/qwen3_5_30b_a3b/test_attention.py``. Unlike qwen's decode step
-(``build_decode_attention``, a fixed-capacity ``cache_update`` write), the
-context length here is a genuine ``DimVar``: the same built Function
-instances evaluate at any concrete prior/total context length without being
-rebuilt (docs/plans/agent-kernel-loop/P0a-tonight-nested-module-e2e.md, M2).
-
-Decode regime: batch 1, exactly one new token per step (``S = 1``, matching
-``moe.py``'s ``(1, 1, DIM)`` token shape) against a dynamic prior context.
+RoPE here uses DeepSeek's interleaved-pairs convention (view-as-complex on
+adjacent dims), distinct from ``tilefoundry.dsl.tf.rope``'s rotate-half
+convention, so it is built from ``tf.slice``/``tf.reshape``/``tf.concat``
+primitives instead of that op.
 """
 from __future__ import annotations
 
-from tests.models.deepseek_v4_flash.moe import DIM
 from tilefoundry import func
-from tilefoundry.dsl import DimVar, ReduceKind, Tensor, tf
+from tilefoundry.dsl import ReduceKind, Tensor, tf
+from tilefoundry.ir.core.module import Module
 
-HEAD_DIM = 128
-NUM_Q_HEADS = 32
-NUM_KV_HEADS = 4
-GQA_GROUP = NUM_Q_HEADS // NUM_KV_HEADS  # 8 query heads share one kv head
-Q_PROJ = NUM_Q_HEADS * HEAD_DIM  # 4096 == DIM
-KV_PROJ = NUM_KV_HEADS * HEAD_DIM  # 512
-
-# One new token decoded per step.
-S = 1
-
-# Decode ceiling for tonight's fixture (M3 exercises ctx 512 and 4096 against
-# the same built Functions — this only bounds the DimVar envelope).
-MAX_CTX = 8192
-
-P = DimVar(name="dsv4_prior_ctx", lo=1, hi=MAX_CTX)
-C = DimVar(name="dsv4_ctx_total", lo=1, hi=MAX_CTX + 1)
-ROPE_CACHE = DimVar(name="dsv4_rope_cache", lo=1, hi=MAX_CTX + 1)
-
-
-@func
-def attn_kv_update(
-    hidden: Tensor[(1, S, DIM), "bf16"],
-    gamma_in: Tensor[(DIM,), "bf16"],
-    w_k: Tensor[(1, DIM, KV_PROJ), "bf16"],
-    w_v: Tensor[(1, DIM, KV_PROJ), "bf16"],
-    gamma_k: Tensor[(HEAD_DIM,), "bf16"],
-    cos_cache: Tensor[(ROPE_CACHE, HEAD_DIM), "bf16"],
-    sin_cache: Tensor[(ROPE_CACHE, HEAD_DIM), "bf16"],
-    pos_ids: Tensor[(S,), "i32"],
-    k_cache_prev: Tensor[(1, P, NUM_KV_HEADS, HEAD_DIM), "bf16"],
-    v_cache_prev: Tensor[(1, P, NUM_KV_HEADS, HEAD_DIM), "bf16"],
-):
-    # Full cache after key RMSNorm + RoPE + append: (k_full, v_full) of shape
-    # [1, prior+1, kv_heads, head_dim] (returned, not matmul'd).
-    hidden_norm = tf.rms_norm(hidden, gamma_in)
-    k = tf.reshape(tf.matmul(hidden_norm, w_k), new_shape=(1, S, NUM_KV_HEADS, HEAD_DIM))
-    v = tf.reshape(tf.matmul(hidden_norm, w_v), new_shape=(1, S, NUM_KV_HEADS, HEAD_DIM))
-    k_norm = tf.rms_norm(k, gamma_k)
-    _, k_rope = tf.rope(k_norm, k_norm, cos_cache, sin_cache, pos_ids)
-    k_full = tf.concat(k_cache_prev, k_rope, axis=1)
-    v_full = tf.concat(v_cache_prev, v, axis=1)
-    return (k_full, v_full)
-
-
-@func
-def attn_scores(
-    hidden: Tensor[(1, S, DIM), "bf16"],
-    gamma_in: Tensor[(DIM,), "bf16"],
-    w_q: Tensor[(1, DIM, Q_PROJ), "bf16"],
-    gamma_q: Tensor[(HEAD_DIM,), "bf16"],
-    cos_cache: Tensor[(ROPE_CACHE, HEAD_DIM), "bf16"],
-    sin_cache: Tensor[(ROPE_CACHE, HEAD_DIM), "bf16"],
-    pos_ids: Tensor[(S,), "i32"],
-    k_full: Tensor[(1, C, NUM_KV_HEADS, HEAD_DIM), "bf16"],
-    v_full: Tensor[(1, C, NUM_KV_HEADS, HEAD_DIM), "bf16"],
-    attn_mask: Tensor[(1, 1, S, C), "bf16"],
-    scale: Tensor[(1, 1, 1, 1), "bf16"],
-    w_o: Tensor[(1, Q_PROJ, DIM), "bf16"],
-) -> Tensor[(1, S, DIM), "bf16"]:
-    # GQA attention of the new token's query over the full (dynamic-length)
-    # context, then the output projection.
-    hidden_norm = tf.rms_norm(hidden, gamma_in)
-    q = tf.reshape(tf.matmul(hidden_norm, w_q), new_shape=(1, S, NUM_Q_HEADS, HEAD_DIM))
-    q_norm = tf.rms_norm(q, gamma_q)
-    q_rope, _ = tf.rope(q_norm, q_norm, cos_cache, sin_cache, pos_ids)
-
-    k_b = tf.repeat_interleave(k_full, repeats=GQA_GROUP, axis=2)
-    v_b = tf.repeat_interleave(v_full, repeats=GQA_GROUP, axis=2)
-    q_h = tf.transpose(q_rope, perm=(0, 2, 1, 3))
-    k_h = tf.transpose(k_b, perm=(0, 2, 1, 3))
-    v_h = tf.transpose(v_b, perm=(0, 2, 1, 3))
-    q_s = tf.mul(q_h, scale)
-    k_t = tf.transpose(k_h, perm=(0, 1, 3, 2))
-    scores = tf.add(tf.matmul(q_s, k_t), attn_mask)
-    probs = tf.softmax(scores, axis=-1)
-    ctx = tf.matmul(probs, v_h)
-    attn_out = tf.transpose(ctx, perm=(0, 2, 1, 3))
-    return tf.matmul(tf.reshape(attn_out, new_shape=(1, S, Q_PROJ)), w_o)
-
-
-# ═════════════════════════════════════════════════════════════════════════
-# Stage 2 (best-effort, new symbols only): the GQA placeholder above is left
-# byte-for-byte untouched — decode_step.py / test_decode_step_e2e.py, owned
-# by another agent working in this worktree, import attn_kv_update /
-# attn_scores / HEAD_DIM / NUM_Q_HEADS / NUM_KV_HEADS / GQA_GROUP /
-# Q_PROJ / KV_PROJ / S / P / C / ROPE_CACHE / MAX_CTX directly.
-#
-# Real-dimension DeepSeek-V4-Flash MLA attention decode (single new token,
-# start_pos > 0 -- matching this file's own existing decode-only scope).
-# Scope: real transformer layer 0 (config.json ``compress_ratios[0] == 0``)
-# -- pure sliding-window MLA: low-rank Q (wq_a -> q_norm -> wq_b), a single
-# shared 512-dim KV latent (wkv -> kv_norm; MQA, n_kv_heads == 1), a FIXED
-# 128-token sliding-window cache (no unbounded growth -> a genuine static
-# shape, no DimVar needed here unlike the placeholder above), attn_sink
-# -augmented softmax, grouped low-rank O projection (wo_a per-group ->
-# wo_b). Cross-checked against this worktree's stage-1 torch oracle,
-# ``tests/models/deepseek_v4_flash/hf_attention_ref.py``'s
-# ``AttentionRef(layer_id=0, ...)``, via ``test_attention_hir_oracle.py``.
-#
-# NOT attempted: real layers with compress_ratio != 0 (the learned Indexer +
-# Compressor top-k KV-compression sparsity -- e.g. layer 2, the layer
-# hf_attention_ref.py's oracle uses for its harder / index_topk coverage).
-# See the run report for the concrete op-contract reasons: TopK's static
-# ``k`` attribute vs. index_topk's context-length-dependent
-# ``min(512, end_pos // 4)``; Gather's ``batch_dims`` convention not
-# matching the per-(batch,query-position)-own-index-set KV gather the
-# Indexer's selection needs; the Compressor's conditional/stateful
-# ring-buffer update (an ``if should_compress: ... else: return None``
-# gating a cache write) not obviously expressible as a pure dataflow
-# Function.
-#
-# ``tilefoundry.dsl.tf.rope`` implements the "rotate_half" (half-split)
-# convention (see its evaluator: ``q*cos + rotate_half(q)*sin``);
-# DeepSeek-V4's official ``apply_rotary_emb`` (hf_attention_ref.py, ported
-# from model.py) uses the "interleaved-pairs" convention (view-as-complex on
-# adjacent pairs (x0,x1),(x2,x3),...). These rotate different dimension
-# pairings and are not interchangeable without a permutation adapter (not
-# attempted here) -- RoPE below is instead built from primitives (strided
-# ``tf.slice`` to de-interleave even/odd elements, ``tf.reshape``+
-# ``tf.concat`` to re-interleave), reproducing the official interleaved-pairs
-# math exactly rather than using the ``tf.rope`` op. Confirmed against
-# ``Slice``'s evaluator (builds a literal python ``slice(begin,end,stride)``
-# per axis, so a stride-2 slice is exactly ``x[offset::2]``).
-# ═════════════════════════════════════════════════════════════════════════
-
+# Real transformer layer-0 dimensions (config.json compress_ratios[0] == 0).
 REAL_DIM = 4096
 REAL_N_HEADS = 64
 REAL_HEAD_DIM = 512
@@ -164,39 +39,21 @@ REAL_WO_A_IN = REAL_N_HEADS * REAL_HEAD_DIM // REAL_O_GROUPS   # 4096
 REAL_WO_A_OUT = REAL_O_GROUPS * REAL_O_LORA_RANK               # 8192
 REAL_NOPE_DIM = REAL_HEAD_DIM - REAL_ROPE_DIM                  # 448
 
-# KV-cache fp8 fake-quant (official: kernel.py's `act_quant(..., inplace=True)`
-# / hf_attention_ref._fake_quant_fp8_block, round_scale=True): per-block
-# absmax -> power-of-2 scale, applied to the non-rope portion of the cached
-# KV latent only (see mla_kv_update_v2 below).
-KV_QUANT_BLOCK = 64                                # official FP8_ACT_BLOCK_SIZE
+# KV-cache fp8 fake-quant: per-block absmax -> power-of-2 scale, applied to
+# the non-rope portion of the cached KV latent only (see mla_kv_update).
+KV_QUANT_BLOCK = 64                                # fp8 quant block size
 KV_QUANT_BLOCKS = REAL_NOPE_DIM // KV_QUANT_BLOCK  # 7
 FP8E4M3_MAX = 448.0            # max finite magnitude representable in e4m3
 FP8E4M3_QUANT_EPS = 1e-4       # amax floor (guards log2(0) on an all-zero block)
 
-
-# NOTE on `@func` bodies: they are parsed from the Python AST at decoration
-# time (`tilefoundry.parser.hir_parser`), not eagerly traced -- a call must
-# resolve to a registered Op or another `@func`/`Module` reference; a plain
-# (undecorated) Python helper function is rejected ("unknown Op name ...").
-# So RoPE (which needs the same interleave/de-interleave math at 3 call
-# sites) and the 8-group `wo_a` projection are inlined/unrolled below rather
-# than factored into a shared helper or a Python `for` loop -- verbose, but
-# every statement is a direct, individually-proven `tf.<op>(...)` call (see
-# qwen3_5_30b_a3b/qwen3_module.py and this file's own GQA placeholder above
-# for precedent on each op's calling convention).
-#
-# `mla_kv_update_v2` and `mla_attend_v2` are two separate `@func`s, chained
-# in Python by the caller (test_attention_hir_oracle.py) via two `evaluate()`
-# calls -- mirroring this file's own placeholder's
-# attn_kv_update -> attn_scores split (see its docstring) -- rather than one
-# `@func` calling the other internally: only *methods on the same `@module`
-# class* calling each other is demonstrated elsewhere (qwen3_module.py); a
-# standalone `@func` calling another standalone `@func` has no such
-# precedent in this codebase, so it is avoided here rather than assumed.
+# RoPE (needed at 3 call sites) and the 8-group wo_a projection are unrolled
+# below via explicit tf.* ops rather than a shared Python helper or a `for`
+# loop: a @func body must resolve every call to a registered Op or another
+# @func, so a plain (undecorated) Python helper is not an option here.
 
 
 @func
-def mla_kv_update_v2(
+def mla_kv_update(
     hidden: Tensor[(1, 1, REAL_DIM), "bf16"],
     gamma_kv: Tensor[(REAL_HEAD_DIM,), "bf16"],
     w_kv: Tensor[(REAL_DIM, REAL_HEAD_DIM), "bf16"],
@@ -259,7 +116,7 @@ def mla_kv_update_v2(
 
 
 @func
-def mla_attend_v2(
+def mla_attend(
     hidden: Tensor[(1, 1, REAL_DIM), "bf16"],
     gamma_q_lora: Tensor[(REAL_Q_LORA_RANK,), "bf16"],
     w_q_a: Tensor[(REAL_DIM, REAL_Q_LORA_RANK), "bf16"],
@@ -399,20 +256,10 @@ def mla_attend_v2(
     return tf.matmul(y, w_o_b)
 
 
+attention_module = Module(name="attention", functions=(mla_kv_update, mla_attend), entry="mla_attend")
+
+
 __all__ = [
-    "C",
-    "GQA_GROUP",
-    "HEAD_DIM",
-    "KV_PROJ",
-    "MAX_CTX",
-    "NUM_KV_HEADS",
-    "NUM_Q_HEADS",
-    "P",
-    "Q_PROJ",
-    "ROPE_CACHE",
-    "S",
-    "attn_kv_update",
-    "attn_scores",
     "REAL_DIM",
     "REAL_HEAD_DIM",
     "REAL_N_HEADS",
@@ -426,6 +273,7 @@ __all__ = [
     "REAL_WINDOW",
     "REAL_WO_A_IN",
     "REAL_WO_A_OUT",
-    "mla_attend_v2",
-    "mla_kv_update_v2",
+    "attention_module",
+    "mla_attend",
+    "mla_kv_update",
 ]
