+++ tests/models/qwen3_5_35b_a3b/config.py
@@ -1,0 +1,6 @@
+"""Qwen3.5-35B-A3B shared model-wide dimensions (per-component dims live in
+their own component file)."""
+from __future__ import annotations
+
+HIDDEN = 2048
+VOCAB = 248320
